amqp: Implicitly create exchanges instead of queues
For the case of making outqueue==inqueue work, we used
to implicitly create queues and bind them to exchange of same name.
This had the problem that whenever someone sent something on the
exchange, it would stay in a queue forever if noone was subscribed to it.

Fixes msgflo/msgflo#10